Removed the 'enabled' flag. Added empty implementation
