+++ controller/internal/sentry/default.go
@@ -1,0 +1,14 @@
+package sentry
+
+import (
+	"context"
+
+	"github.com/getsentry/sentry-go"
+)
+
+type Default struct {
+}
+
+func (s *Default) Capture(ctx context.Context, err error) {
+	sentry.CaptureException(err)
+}
