+++ controller/internal/sentry/disabled.go
@@ -1,0 +1,12 @@
+package sentry
+
+import (
+	"context"
+)
+
+type Disabled struct {
+}
+
+func (d *Disabled) Capture(ctx context.Context, err error) {
+	// This implementation does nothing.
+}
