+++ controller/internal/sentry/spec.go
@@ -1,0 +1,10 @@
+package sentry
+
+import (
+	"context"
+)
+
+type Interface interface {
+	// Capture is used to send the error to Sentry API.
+	Capture(ctx context.Context, err error)
+}
